Rename me/profile directory to me/account
